Rework FlixelBatch to be an interface for extensibility
