+++ flixelgdx-core/src/main/java/org/flixelgdx/util/timer/package-info.java
@@ -9,7 +9,7 @@
  *   <li>Schedule work: {@link org.flixelgdx.util.timer.FlixelTimer#getGlobalManager()}{@code .start(seconds, callback, loops)}
  *     or static helpers {@link org.flixelgdx.util.timer.FlixelTimer#wait(float, org.flixelgdx.util.timer.FlixelTimerListener)}
  *     and {@link org.flixelgdx.util.timer.FlixelTimer#loop(float, org.flixelgdx.util.timer.FlixelTimerListener, int)}.</li>
- *   <li>Scaling: {@link org.flixelgdx.FlixelGame} passes {@code elapsed} times {@link org.flixelgdx.Flixel#getTimeScale()} into
+ *   <li>Scaling: {@link org.flixelgdx.FlixelGame} passes {@code elapsed} times {@link org.flixelgdx.Flixel#timeScale()} into
  *     {@link org.flixelgdx.util.timer.FlixelTimerManager#update(float)}.</li>
  *   <li>Pooled instances: do not store {@link org.flixelgdx.util.timer.FlixelTimer} references across {@link org.flixelgdx.util.timer.FlixelTimer#cancel()} or completion;
  *     the manager returns them to an internal {@link com.badlogic.gdx.utils.Pool}.</li>
